backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.0
timm==0.9.12
scikit-learn==1.3.0
h5py==3.9.0
transformers==4.37.2
tqdm==4.67.1
python-dotenv==1.1.0